changes based on code review: clearing interval, only setting interval when helpscreen showing, moved files to reactViewModels, updateCurrentRectangle moved
