+++ lib/ReactViewModels/HelpSequence.js
