+++ lib/ReactViewModels/HelpSequences.js
